:bulb: impl internal_cls
